[ceph] split the ceph plugin
This work distributes the ceph plugin into plugins
for individual ceph components (mon,osd,mds,rgw,mgr),
so that additional data collection can then be added
in each of the components. Work for additional data
collection in each component plugin will follow in
a later commit.

Closes: #1945
